First draft of eratosthenes w/ wheel [CI SKIP]

+++ include/boost/math/special_functions/detail/interval_prime_sieve.hpp
@@ -185,7 +185,7 @@ void IntervalSieve<Integer, OutputIterator>::Sieve() noexcept
     
     else
     {
-        primes_.resize(prime_approximation(right_));
+        primes_.resize(static_cast<std::size_t>(prime_approximation(right_)));
         linear_sieve(primes_range, primes_.begin());
 
         for(; primes_[i] < primes_range; ++i)
